Migrate to TypeScript and refactor repo structure
