--- .eslintignore
@@ -1,2 +1,0 @@
-react/node_modules/*
-react/coverage/*
