+++ package.json
@@ -1,56 +1,15 @@
 {
-  "version": "2.0.0",
   "devDependencies": {
     "@vtex/intl-equalizer": "^2.0.1",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "7",
-    "babel-jest": "^23.6.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-app": "^3.1.1",
-    "eslint": "^5.10.0",
-    "eslint-config-prettier": "^3.3.0",
-    "eslint-config-vtex": "^8.0.0",
-    "eslint-config-vtex-react": "^3.0.3",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-lodash": "^5.1.0",
-    "eslint-plugin-prettier": "^3.0.0",
-    "husky": "^1.3.0",
-    "jest": "^22.0.4",
-    "prettier": "^1.15.3",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-intl": "^2.7.2",
-    "react-testing-library": "^5.4.2",
-    "regenerator-runtime": "^0.13.1"
+    "husky": "^1.3.1"
   },
   "scripts": {
-    "lint": "eslint react/**/*.js",
-    "lint:locales": "intl-equalizer",
-    "test": "jest",
-    "test:watch": "jest --watch"
+    "lint:messages": "intl-equalizer"
   },
   "husky": {
     "hooks": {
-      "pre-push": "yarn run lint && yarn run test && yarn run lint:locales"
+      "pre-commit": "cd ./react && yarn lint",
+      "pre-push": "yarn lint:messages && cd ./react && yarn test && yarn lint"
     }
-  },
-  "jest": {
-    "collectCoverageFrom": [
-      "react/**/*.{js,jsx}"
-    ],
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "react/node_modules/",
-      "react/__mocks__/"
-    ],
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs)$"
-    ],
-    "testMatch": [
-      "<rootDir>/react/**/__tests__ /**/*.{js,jsx,ts,tsx}",
-      "<rootDir>/react/**/?(*.)(spec|test).{js,jsx,ts,tsx}"
-    ],
-    "testURL": "http://localhost/"
   }
 }
