+++ react/package.json
@@ -1,8 +1,58 @@
 {
   "name": "shippint-estimate-translator",
-  "version": "0",
   "description": "",
   "dependencies": {
+    "react": "^16.8.1",
     "react-intl": "^2.7.2"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.2.2",
+    "@babel/preset-env": "^7.3.1",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.1.0",
+    "@types/graphql": "^14.0.5",
+    "@types/jest": "^24.0.0",
+    "@types/node": "^10.12.24",
+    "@types/prop-types": "^15.5.8",
+    "@types/react": "^16.8.2",
+    "@types/react-intl": "^2.3.16",
+    "babel-jest": "^24.1.0",
+    "jest": "^24.1.0",
+    "react-dom": "^16.8.1",
+    "react-testing-library": "^5.6.0",
+    "regenerator-runtime": "^0.13.1",
+    "tslint": "^5.12.1",
+    "tslint-config-vtex": "^2.1.0",
+    "tslint-eslint-rules": "^5.4.0",
+    "typescript": "^3.3.3",
+    "typescript-babel-jest": "^1.0.6"
+  },
+  "resolutions": {
+    "babel-core": "^7.0.0-bridge.0"
+  },
+  "scripts": {
+    "lint": "eslint react/**/*.js",
+    "lint:locales": "intl-equalizer",
+    "test": "jest",
+    "test:watch": "jest --watch"
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "react/**/*.{ts, tsx}"
+    ],
+    "transform": {
+      "^.+\\.(js|jsx|mjs)$": "<rootDir>/node_modules/babel-jest",
+      "^.+\\.(ts|tsx)$": "typescript-babel-jest"
+    },
+    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$",
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx",
+      "json",
+      "node",
+      "graphql"
+    ]
   }
 }
