--- react/shapes/index.js
@@ -1,6 +1,0 @@
-import PropTypes from 'prop-types'
-
-export const deliveryWindowShape = PropTypes.shape({
-  startDateUtc: PropTypes.string,
-  endDateUtc: PropTypes.string,
-})
