+++ react/test-utils/index.tsx
@@ -1,9 +1,9 @@
-import React from 'react'
+import React, { ReactNode } from 'react'
 import { IntlProvider } from 'react-intl'
 import { render } from 'react-testing-library'
-import messages from '../messages/pt.json'
+import messages from '../../messages/pt.json'
 
-const renderWithIntl = node => {
+const renderWithIntl = (node: ReactNode) => {
   return render(
     <IntlProvider locale="pt" messages={messages}>
       {node}
