+++ react/tslint.json
@@ -1,0 +1,6 @@
+{
+  "extends": "tslint-config-vtex",
+  "rules": {
+    "object-literal-sort-keys": false
+  }
+}
