+++ react/typings/global.d.ts
@@ -1,0 +1,4 @@
+interface DeliveryWindow {
+  startDateUtc: string | null
+  endDateUtc: string | null
+}
